Fix crash when saving override config with missing parent directory
Summary: `/var/opt/magma/configs` is not created when we spin up the VM initially, setting the log level using `config_cli` would crash in this case because it cannot save the file in a non-existent directory. This diff fixes the issue by creating the directory if necessary before saving file.

Reviewed By: andreilee

Differential Revision: D14714326

fbshipit-source-id: 83398c91a9dc8907292f070c5689f76e05a6ec30