Add docker init system (#154)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/154

This change adds an option to use docker as an init system as an alternative to systemd and runit. This requires docker to be installed, and if running in a container, the docker socket should be mounted. This new init system will be used to dockerize the FeG and run one container per service. Then, docker will be used to manage the services instead of systemd.

Reviewed By: alexanderclin

Differential Revision: D14687148

fbshipit-source-id: cc567711ca8b772c2904ebe175f2e33a9685ddcd