Fixed access gateway documentation
Summary:
In the process of building the docs site, some READMEs for Access Gateway were
moved, leading to broken symlinks. This change fixes it.

Additionally, this change also makes edits to the call flow diagram, which was
missing an S1AP state and call to openflow_add_tunnel from SGW task

bypass-lint

Reviewed By: amarpad

Differential Revision: D14729418

fbshipit-source-id: 8828c9a6e156f6d90585dac6e72e2781115b2ab0