Change ServiceRegistry to read in 'proxy_aliases' fields from service_registry.yml files (#153)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/153

Currently there are multiple lists of services, when the service_registry YAML files should be the only place and everything else should point to the registry. My goal is to get rid of `magma/orc8r/cloud/deploy/files/envdir/CONTROLLER_SERVICES`, which currently lists controller services.
This is step one in trying to achieve this goal. Right now, not all of the fields in service_registry.yml files are being loaded into the corresponding struct in Go.

ServiceRegistry now reads in the field proxy_aliases which was added back in D14176660, so these can be used to specify different proxy ports.

Reviewed By: fishlinghu

Differential Revision: D14428141

fbshipit-source-id: 47fd491e2d4e0a1ff2069d5848362e22f9a09c56