Add s1ap_state get/put + mock redis store/load
Summary:
As part of our goal to make MME stateless (can crash/update and come back up without losing enb/ue context), we are making the S1AP task stateless as a POC. To make a C program "stateless", we are going off the idea that only the database state needs to be good at all times -- therefore we let the message/state processing happen as normal except that the state is moved from separate globals into one global state that is committed to the db at the end of processing each message.

In this diff, we add the stateless S1AP state type (s1ap_state_t) and what are essentially mock methods to init, get, put, and exit the stateless S1AP context.

Reviewed By: ssanadhya

Differential Revision: D14565849

fbshipit-source-id: 9b6776791d69cc22861c8e11a037dc5827919a2b