Change s1ap task to use s1ap state instead of ht globals
Summary:
As part of our goal to make MME stateless (can crash/update and come back up without losing enb/ue context), we are making the S1AP task stateless as a POC. To make a C program "stateless", we are going off the idea that only the database state needs to be good at all times -- therefore we let the message/state processing happen as normal except that the state is moved from separate globals into one global state that is committed to the db at the end of processing each message.

In this diff, we change the S1AP state management from separate globals to one global state. We then use the s1ap_state mocks to get and put the state (also init and exit when necessary).

Reviewed By: ssanadhya

Differential Revision: D14565851

fbshipit-source-id: ad40d821215038b4bba2cb48def9316fe370f49c