Add s1ap_state get enb/ue funcs
Summary:
As part of our goal to make MME stateless (can crash/update and come back up without losing enb/ue context), we are making the S1AP task stateless as a POC. To make a C program "stateless", we are going off the idea that only the database state needs to be good at all times -- therefore we let the message/state processing happen as normal except that the state is moved from separate globals into one global state that is committed to the db at the end of processing each message.

In this diff, we add getter functions for several parts of S1AP task state (get enb and ue). The purpose of this is to mask implementation detail about how ue's and enb's are stored so that we can change the implementation from OAI hashtable's with manual memory management to C++ maps in the future.

Reviewed By: ssanadhya

Differential Revision: D14685279

fbshipit-source-id: 7b99a936bab0af073bba91cc4c59dbbf626fd7f7