Add to/from protos for eNB and UE
Summary:
As part of our goal to make MME stateless (can crash/update and come back up without losing enb/ue context), we are making the S1AP task stateless as a POC. To make a C program "stateless", we are going off the idea that only the database state needs to be good at all times -- therefore we let the message/state processing happen as normal except that the state is moved from separate globals into one global state that is committed to the db at the end of processing each message.

In this diff, we add protos for eNB/UE description and S1AP state. Additionally, we implement to and from methods for each of the protos to convert between them and their corresponding c structs.

Reviewed By: ssanadhya

Differential Revision: D14581888

fbshipit-source-id: cd3050d0a8e564c5727f9b9309152b38227ac920