Convert lookup table to _2str function
Summary:
Currently, for some logging, we rely on lookup tables without bounds checking or guarantees about enum values to convert certain values to strings. This could cause MME crashes or incorrect string values in the future.

In this diff, we change this from lookup tables to functions with switch/case statements.

Reviewed By: vikg-fb

Differential Revision: D14690806

fbshipit-source-id: 04518dcda0ddb64d1bb99ebd0e5114027897f232