Use s1ap_state get enb/ue funcs and remove s1ap_mme versions
Summary:
As part of our goal to make MME stateless (can crash/update and come back up without losing enb/ue context), we are making the S1AP task stateless as a POC. To make a C program "stateless", we are going off the idea that only the database state needs to be good at all times -- therefore we let the message/state processing happen as normal except that the state is moved from separate globals into one global state that is committed to the db at the end of processing each message.

In this diff, we change s1ap task to use the s1ap_state getter methods introduced in the previous diff. We also remove the old methods provided by s1ap_mme.c that did the same thing.

Reviewed By: ssanadhya

Differential Revision: D14685282

fbshipit-source-id: a34ee5da7bb9a4fd35c904a6081afcac48452534