Asynchronously terminate session for subscriber
Summary:
**Background:**
The initial implementation of PCEF simplified the flow deletion process by terminating the session immediately upon receiving an EndSession request. This resulted in under-reporting of flow usage stats, because pipelined polls flow stats periodically and any usage between the last poll and the time of termination is lost.

D14116606 modifies pipelined to schedule deletion after the last usage us reported, thus fixing the under-reporting problem for the service. However, currently, sessiond still terminates the session immediately. As a result, when pipelined reports the last usage for the session, the usage is ignored since the session no longer exists in sessiond.

**Implementation:**
This diff modifies sessiond to terminate the session asynchronously, so it will wait for the last report from pipelined. With this diff, the entire under-reporting problem will be solved on a system level.

We introduce additional states for a session to manage its lifecycle:
```
  /**
   * State transitions of a session:
   * SESSION_ACTIVE
   *       |
   *       | (start_termination)
   *       V
   * SESSION_TERMINATING_FLOW_ACTIVE <----------
   *       |                                   | (add_used_credit)
   *       | (new_report)                      |
   *       V                                   |
   * SESSION_TERMINATING_AGGREGATING_STATS -----
   *       |
   *       | (finish_report)
   *       V
   * SESSION_TERMINATING_FLOW_DELETED
   *       |
   *       | (complete_termination)
   *       V
   * SESSION_TERMINATED
   */
```
When an EndSession request is received,  `start_termination` is called to mark the subscriber session to be terminated. sessiond will delay the actual session termination until a usage report that does not contain the subscriber, which means that pipelined has deleted the flow and finished reporting its usage. (Technically speaking, we can terminate the session sooner by implementing an endpoint or modifying the RuleRecord proto to specify that usage reporting is finished, but we can keep it simple for now and just wait for the next report) Once this happens, a request along with the last usage stats is sent to Cloud to terminate the session, and the session is deleted locally.

A force termination timeout is included in the config YAML. In the event that flow deletion fails, the session will be terminated forcefully after the timeout. This is implemented by scheduling a `complete_termination` call in `start_termination`.

See header comments for more details.

**Notable changes**
- We will not surface termination errors to MME, since termination is done asynchronously now. This is fine since we do not do any error handling in MME anyways right now. If we want to do error handling, then a new endpoint will need to be introduced in MME for sessiond to surface the error.
- Local termination no longer waits for Cloud to finish. Previous implementation performed local termination after Cloud finishes, no matter whether Cloud termination succeeds or not, so there should be no functional changes. Maintaining the existing behavior would make the code even more complicated as we are making the local termination asynchronous now.

Reviewed By: amarpad

Differential Revision: D14669346

fbshipit-source-id: a0885e68e80ae0406630c71a0bca83ba25325dd9