Move readmes into the readmes folder
Summary: Some READMEs are missing from the `readmes/` folder. This change adds all of them to the folder and adds symlinks to them throughout the repo where they are needed. Also, the FeG readme is broken into 3 smaller readme's so that it can be linked to from the sidebar in docusaurus more conveniently.

Differential Revision: D14783507

fbshipit-source-id: 1a7a1a156389bb40e7fe3b4ccb19efecc04f9c64