Move readmes into the readmes folder (#169)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/169

Some READMEs are missing from the `readmes/` folder. This change adds all of them to the folder and adds symlinks to them throughout the repo where they are needed. Also, the FeG readme is broken into 3 smaller readme's so that it can be linked to from the sidebar in docusaurus more conveniently.

Reviewed By: alexanderclin

Differential Revision: D14783507

fbshipit-source-id: d98265b5fcfcebaa79778ed2490b69192060fb28