Dockerize FeG
Summary:
**Summary**

This change enables running the FeG using docker instead of using vagrant and ansible. The FeG can be run either for development or for production. In development mode, all FeG services are run in a single docker container. This allows for a very similar development workflow compared to using a VM. In production mode, each FeG service is run in its own container and docker is used for service management. The production containers have been optimized for size to ease deployment and improve isolation.

**Implementation**

Two new docker files were created and stored in `feg/gateway/build`. They are for go and python. Then, a docker-compose file was made for the development setup and the production setup. This stores all the details needed to create docker containers out of the go and python images. In production, the magmad container mounts the docker socket and a docker init system option was added to magmad's service manager. This allows docker to manage the services instead of requiring systemd to be used.

Reviewed By: alexsn, mpgermano

Differential Revision: D14586144

fbshipit-source-id: 708ed177d12fd0741fa2d7800d37cf83b62469b2