Add snapshot testing support for pipelined tests
Summary:
**Background:**
Currently, when writing unit tests for pipelined, we rely on sending a flow stats request to OVS to check the presence of a flow. This approach has 2 issues:
- Actions cannot be verified as it is not included in the reply
- Forgetting a match field in the query will usually still return the correct flow, making it easy to leave out fields when verifying the flows

In general, it is hard to write unit tests to catch changes and regressions in the flows installed by pipelined.

**Implementation**
This diff introduces snapshot testing to test installed flows by pipelined. The snapshots are annotated flow dumps from OVS without non-deterministic fields such as `duration` and `idle_age`. They will be saved as files inside `tests/snapshots`.

In tests, `assert_bridge_snapshot_match` can be used to assert that the flow installed in the test is unchanged from existing snapshots. If a mismatch occurs, the test will fail with the diff in snapshots and the new snapshot.

We will check in the snapshots into source control, so it will be clear to track any changes and side effects to flow installation.

Reviewed By: amarpad

Differential Revision: D14714608

fbshipit-source-id: 1019a8da7dd578866d45ea41d5d9120debb472b1