Refactoring service_poller logic to use Job interface / async model (#152)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/152

Small refactoring to make the code a little bit easier to read. It now implements a Job interface defined in D14006075. So instead of implementing the loop logic in service_poller, it now only implements the work it needs to do and a loop interval, and relies on the Job interface to periodically execute the work.

Reviewed By: xjtian

Differential Revision: D14685028

fbshipit-source-id: ef88a3e8aabb5d6cd35bfce97c32fae7a088e5e5