Cleanup the proxy ports and make orc8r docker work for AGW
Summary:
There is a lack of clarity in the ports using by the controller and the nms.
We will be using 9443 and 9444 for the controller, and 443 for the nms.

Also with this change, the AGW would be able to speak seamlessly to the vagrant or docker instances since
they bind to the same port in the host and port forward.

Reviewed By: xjtian

Differential Revision: D14863967

fbshipit-source-id: 54ba884cf340d2e272fbf8a3dd2cf8a96ec043fe