Graphite query swagger definitions
Summary:
Create swagger definitions for querying the graphite database. The `magma/networks/:network_id/graphite/query` will be the endpoint used to request data from graphite.
This is a temporary solution while I work on multi-tenant support so functions will not be able to be used yet. Query should be of the form `<metric_name>[;tag=tagValue]*`

Reviewed By: xjtian

Differential Revision: D14820311

fbshipit-source-id: c316ddb9a75127ae962e558039c46460012c0150