graphite query handlers (#183)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/183

Implements the GET handler for graphite query requests. This will allow the switch from prometheus to graphite as metrics will be received in the database and are now accessible to users. Basic multi-tenant support is in place by validating networkID and wrapping the query in a function that restricts by this tag.
Also moves common functionality between prometheus and graphite handlers into own util package.

Reviewed By: xjtian

Differential Revision: D14876647

fbshipit-source-id: 35a48cb68c48c67d5469f76ed83f59a40ac5249f