Support server-streaming within Sync RPC (#139)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/139

This diff updates Sync RPC to support server-streaming. This requires some changes to the Sync RPC protobuf, sync rpc client, proxy client, and dispatcher.

The dispatcher HTTP server has been updated to support multiple responses by wrappping the `select` statement in `rootHandler` with a loop. It only returns upon timeout or when `Grpc-Status` is seen in the Trailer. In addition, when the request context is done, a message is sent down to the gateway to let the proxy client know to stop receiving frames for this particular request ID.

The Sync RPC client maintains a table that maps request IDs to bools that dictates whether the connection for this particular request ID is closed.

The proxy client now sends responses frame by frame rather than waiting for all frames. When there are no more frames to receive the proxy client returns.

`sync_rpc_service`.proto has been updated with new fields `keepConnActive` and `connClosed`:
```
message GatewayResponse {
    ...
    bool keepConnActive = 5;
}

message SyncRPCRequest {
    ...
    bool connClosed = 4;
}
```
`keepConnActive` is used by the proxy client to tell the cloud client that the connection is still alive.
`connClosed` is sent from the cloud to the gateway to signify that the proxy client should stop reading frames.

Reviewed By: rpraveen

Differential Revision: D14541497

fbshipit-source-id: 3bc37acc8052e6889fc6817a3cd3106c29c1921c