Custom Cavium GetObjects
Summary:
Using get_all_objects_to_add to get the list of objects to add results in a Fault if the objects don't exist. Right now, we are using this for getting the PLMN list. On Cavium, this list is initially empty. Instead, we get the parent PLMN List object which will include all the children if they exist.

Adding CaviumWaitGetObjectParametersState which does nothing but cleans up the constructor since we always have to disable admin state before trying to change the configs.

Differential Revision: D14602266

fbshipit-source-id: 75e366776a2eb7b48c8e3f09b85f7e1e6d2f713f