GetRPCMethods model
Summary:
The current GetRPCMethod handler sends a malformed response to the Cavium, failing to correctly form the MethodList.
Adds the full response model as well as corresponding state. This state can be used to track when we have a reboot since GetRPCMethods only occurs after Inform message after a reboot/loss of eNB state.

Differential Revision: D14599889

fbshipit-source-id: 062b81847d45c8f6e886828a5259e7e6ada9955b