Integrate docusaurus bot
Summary: Currently the docs and the documentation website can get out of sync if the website is not updated. To prevent this from happening, the docusaurus bot will now automatically update the documentation website when the docs are changed. Docs will be stored on the gh-pages branch instead of inside the docs/web folder since this is required by the bot. A future change will remove the docs/web folder once we have switched to gh-pages.

Reviewed By: rpraveen, sciencemanx

Differential Revision: D14937365

fbshipit-source-id: 1a8950fdcda90d77ce6f26ca394124240603c991