GetRPCMethods model (#163)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/163

The current GetRPCMethod handler sends a malformed response to the Cavium, failing to correctly form the MethodList.
Adds the full response model as well as corresponding state. This state can be used to track when we have a reboot since GetRPCMethods only occurs after Inform message after a reboot/loss of eNB state.

Reviewed By: andreilee

Differential Revision: D14599889

fbshipit-source-id: 43644066b7daf2f8b488f3a40890a58216ef7a46