Cavium Admin Enable/Disable state
Summary:
In the Cavium device, before we can write any configs, we have to set AdminState to disable. After we are done, we should set it to enable to allow the eNB to start transmitting.
This change extends the relevant classes to support the enabling the admin state.
This adds a check to see if we need to actually apply configurations before disabling AdminState and bringing down the cell. Also checks that we are allowed to enable the cell from mconfig before re-enabling. Includes relevant state transitions.

Reviewed By: andreilee

Differential Revision: D14601830

fbshipit-source-id: 55f73deee6997ecbc543d2a557b0fb9eee5d08d8