link metric sequelize model, initial seed
Summary:
This allows us to move the link_metrics (snr, mcs, rssi, etc) from code in bqs to the DB, so we can dynamically add new metrics as needed.
I started working on the UI but fixing react-router in NMS config is going to take longer than expected so just this for now.

- Model for link_metric
- Initial seed, bqs support in D14587394

Reviewed By: aclave1, csmodlin

Differential Revision: D14679474

fbshipit-source-id: e89488cb1e67593692341df29f481bb8c7257e7e