Make graphite attempt to reconnect after error submitting metric
Summary: If the carbon daemon is stopped on the metrics instance, the connection will be closed and metricsd would need to be restarted to get the connection again. This just tells the exporter that the connection is closed once an error occurs. Then on the next export it will try to reconnect before exporting.

Reviewed By: tcirstea

Differential Revision: D14966796

fbshipit-source-id: 99183d85fbb5545bda4928446efd729bbc3429fd