Move swx proxy cache unit test to same folder as cache (#200)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/200

The recent addition of the swx_proxy's cache put it's unit test in a seperate directory.
To remain consistent with the style of the FeG and golang in general, this diff moves the unit
test to the same directory that the cache is defined in, renaming the package to `cache_test`.

Reviewed By: vikg-fb

Differential Revision: D14968815

fbshipit-source-id: 7bb713268ef72f920e858cba881c87cdac95b1c5