Properly handle cloud names and labels in metricsd servicer
Summary:
The MetricsContext object is supposed to hold all the relevant information about a metric. However, with cloud metrics there was ambiguity about the actual networkID and gatewayID due to the different ways that a cloud metric could be named/labeled. This diff makes the context the "source of truth" so exporters do not have to deal with figuring out IDs.

* Parse the network and gateway IDs out of metrics in this order:
  * Inside the metric name: e.g. `metric_name_gatewayId_<gateway>_networkId_<network>`
  * From the labels of the metric
  * Default values `networkID: "cloud"`, `gatewayID: <hostname>`
* Refactor a lot of the testing code to reuse constants
* Add more complete unit tests to servicer, graphite
* Drop `networkId` and `gatewayId` labels in the graphite exporter since those will be parsed into the context

Differential Revision: D14944924

fbshipit-source-id: 87ab9d7e6a683209637cb18567428d305196c735