Docker ansible role to run docker containers in production
Summary:
Ansible role to upgrade a docker compose based solution to a single VM.
This should be reusable for orc8r, feg and the pkg repo.

I'll change fab package to build a codedeploy zip file with a specific
version in the 'image_version' file and then we can use codedeploy to
deploy a specific version to staging or prod.

Also, this is backward compatible with the old packaging, and so
we can redeploy the old package to go back.

Reviewed By: xjtian

Differential Revision: D15023486

fbshipit-source-id: 03c646aa983b958a4f8fb6276083388adb25d0a7