Refactor docker-compose.prod.yml into generic go and python services
Summary:
This diff serves to simplify the Docker compose file that will be used for the
FeG in production. This diff ensures each service is made up of the same configuration
settings except for any necessary overrides or extra configuration. This will help
to ensure that changes to these settings only need to be made once to apply to all go
or python services.

Reviewed By: vikg-fb

Differential Revision: D15034798

fbshipit-source-id: 84977d114e13403d86bd61effd508b1a52ca4bbc