State machine unhandled errors will have stacktrace logged, will enter failure state
Summary:
This revision makes several changes to aid in debugging of enodebd and eNB behavior.

Unhandled errors in the state machine will trigger a state transition to an error state, so that the same errors are not raised repeatedly, and the logs only contain the relevant errors.

While the error is caught, the ```traceback``` module is used to dump the stacktrace out to error logs.

Finally, some debugging info will be dumped into error logs, of the device and desired configuration.

Reviewed By: sciencemanx

Differential Revision: D14954819

fbshipit-source-id: 34bd003a8bed00a580582e2ae150390ac27c22f8