Implement subscriber streaming (#218)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/218

This allows subscribers to be configured for the HSS lite using swagger. This makes test setup easier so that scripts or the HSS subscriber CLI are not needed to populate subscribers. Subscriber streaming is implemented by adding a new listener which waits for updates from the subscriberdb cloud service and then updates the hss subscriber store accordingly. This feature can be enabled/disabled via swagger.

Reviewed By: fishlinghu

Differential Revision: D15038554

fbshipit-source-id: a0174c0c51f19342ea76bbd2f137e5e579f1483b