Make StatsManager into an observer of the state machine (#222)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/222

This revision reverses the relationship between the state machine and StatsManager.

Instead of the state machine (henceforth called the 'handler') being responsible for clearing stats on the StatsManager whenever the eNodeB stops transmitting, the StatsManager will keep a reference to the handler.

The StatsManager will indirectly observe the state of the eNodeB through the handler. When it detects that that the eNodeB has stopped transmitting, then stats will be cleared.

This change is useful for two reasons:
1) This reorganizes dependencies into a stricter hierarchy, and avoids circular dependencies later on.
2) The StatsManager and the handler/state-machine have a better separation of responsibility.

Reviewed By: ssanadhya

Differential Revision: D14663225

fbshipit-source-id: 0c5b30fe0b8745a61e32ad4e37862a9b650edec9