(1/3) Integrate enb acs manager: StatsManager changed to use eNodeB ACS Manager (#224)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/224

StatsManager will have a reference to the eNodeB ACS Manager, and regularly check the device configurations to track stats. Prior to this change, StatsManager would have a reference to a single eNodeB handler/state-machine, but since support is being added to have multiple eNB configured per gateway, this change is necessary.

Reviewed By: xjtian

Differential Revision: D14709619

fbshipit-source-id: a6168794ae04c036033bb4a41fa2f4a6c0f7f531