Add logic to periodically collect operational states from gateway services (#167)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/167

A new service303 function GetOperationalStates was added in D14738655. A new script in magmad, state_reporter, will query each state for its operational states, aggregate them, and send them off to the state service. Since the operational states can look different for every device, it is encoded as a JSON.

Each service can register a callback function to write in the state when GetOperationalStates is called by the state_reporter. The default in every service is None. This is the same model that the service_poller uses.

Additionally, the reporter will also take the latest checkin request sent as a state and report it.

Reviewed By: xjtian

Differential Revision: D14754977

fbshipit-source-id: af01a62b881f88e589fdb0a87d243e49cc8c55c1