Turn download into an orchestrator service (#137)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/137

Pull Request resolved: https://github.com/facebookexternal/magma/pull/180

This diff makes the download service into a OrchestratorService on top of what it already does. Additionally, I've moved some of the hardcoded configuration into download.yml so that it can be modified more easily.

The motivation for this diff is similar to D14621048. I would eventually like to get rid of CONTROLLER_SERVICES and just pull everything from the service registry files.

Reviewed By: xjtian

Differential Revision: D14432946

fbshipit-source-id: 2303b599c783046ae4dbffdd45c21507fd35306b