Adjust MME name encoding approach according to hcore requirement
Summary:
**Summary**
H-core MSC uses a MME name format that is different from the format used by TeraVM.
TeraVM does not want the `.` in MME name while H-core MSC needs the `.`.
Also, H-core MSC needs `0x00` as the ending character.
In the future, we plan to construct MME name completely in FeG instead of getting the information from Access Gateway.

**Implementation**
- Change the MME name encoding function in FeG

**What is affected**
TeraVM combine attach will fail.
We need to try to modify the test to use this new format for MME name.

Differential Revision: D15246634

fbshipit-source-id: 1a29a876826769aad54c0b198eb050643b891d89