Modifiy type of gateway state sent from magmad (#255)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/255

Changing the type of state sent from gateway to use a named tuple. This way, we can send the states as a JSON that can easily be decoded into the swagger generated GatewayStatus struct.
I've duplicated a lot of the code here that constructs the checkin request to make this happen. After we fully migrate checkind state reporting/reading to use the state service, we can remove much of the old code.

Reviewed By: xjtian

Differential Revision: D15204266

fbshipit-source-id: 010a92a310ea67ffb830f32b10e74033d32783b3