Add a gatewaystatus REST API endpoint for state
Summary:
Added a REST endpoint to get gateway state from the state service.
When the GET request is made, the handler reads from the state storage and returns a serialized GatewayStatus. The result is identical to what the checkind service returns.

In order to get this to work, I had to copy over the swagger model for gateway status from checkind. The other option was to have the model defined higher up in scope, which seemed like a bit of an overkill, so I will leave it like this for now.

{F157345033}

Reviewed By: xjtian

Differential Revision: D15174514

fbshipit-source-id: 9085b902881ab5a97af726a708a5800d8806cfb6