Add certificate epiry and timestamp to saved states to support backward compatibility for GatewayStatus (#256)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/256

The motivation for this change is to support backwards compatibility for when state service replaces checkind service. Right now, the checkind API provides two extra fields (time at which the state was received and certificate expiry date) that are added onto CheckinRequest struct to create GatewayStatus.

The state service will now save those two info for every state that it retrieves.

Reviewed By: xjtian

Differential Revision: D15171547

fbshipit-source-id: f6b65195252a11d52b5be6f90d077bca4dcf7707