Implement Deregister GRPC method for SWx Proxy
Summary:
This diff serves to implement the Deregister method for swx_proxy. When a user is
authenticated with an HSS, the AAA server sends a registration request to the HSS that
links the user to the AAA server. If the AAA server ever needs to remove that user,
it should send a USER_DEREGISTRATION SAR request to the HSS. This diff implements
that functionality. While it is unclear what the exact use-case will be for this method,
we will need it at some point in the future.

Reviewed By: fishlinghu

Differential Revision: D15291496

fbshipit-source-id: a5d4f15fc40a4c09a19006fd338c181722eda094