Add support for eNB configuration
Summary:
Prior to this, configuration for an eNodeB is associated directly to a Magma access gateway. In order to support multiple eNodeB's per AGw, we need the ability to configure multiple eNodeB's separately for AGw.

The plan here is to be able add eNodeB configuration to a network, with the key being the eNodeB serial ID, which has been found to be a unique identifier of the hardware in question. A separate tab is added for eNodeB configs. Then it should be possible to move the eNB between AGw's.

To associate the eNB to the AGw, LTE->RAN configs will have a section for associating eNBs by serial ID to the AGw. Associating the eNB to the AGw is useful because then we can detect if the eNB is unexpectedly disconnected/powered-off, otherwise we wouldn't know if there's supposed to be 2 or 3 eNB's attached to the AGw.

Since the eNodeB configuration has redundant fields vs. the other configuration, the plan is to have a transition period where the old config format is deprecated.

This revision adds the ability to create/read/update/delete eNodeB configuration for a network, but not yet the ability to associate it an AGw.

Reviewed By: rckclmbr

Differential Revision: D15047777

fbshipit-source-id: 33efd70a90bbb7108cb654c54a9cb360c783880d