Explicitly define the environment variables in compose file
Summary:
By explicitly defining the variables in compose file, the knobs for Orc8r are clearly documented.
We should only expose variables which we want to result in behavioral change for open source Magma.

We will remove envdir, and use env_file confguration in production. Until then we have the
envdir still as an option to ensure backwards compatibility.

Reviewed By: xjtian

Differential Revision: D15349026

fbshipit-source-id: 522ddfebcda5329650308a97bb440177cba5f8e8