Implement Northbound Configurator servicers (Load, CreateAndUpdate for Network/NetworkEntity) (#297)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/297

This diff implements the northbound portion of servicer functions defined in D15091643 .
The major chunk of this diff (configurator_helper.go) is the translation between the protobuf structs used for grpc and the structs used for storage.

As mentioned in https://fb.quip.com/SDsNADhTn7wj, the northbound servicers are meant for user interaction with the service. So an upcoming diff will implement the REST API portion that uses these client_api functions also defined in this diff.

Validations for the values passed in will come soon.

Delete functions will come in the next diff.

Reviewed By: xjtian

Differential Revision: D15286443

fbshipit-source-id: 9d1265dafc0227e678642dc12198f6af617f910b