Remove builder intermediate image from go Dockerfile
Summary:
This diff serves to remove the builder image used as an intermediate.
This ensures that the test container comes up with the binaries already built
in the container. This helps to speed up development since we don't need to
run `make run` inside the container in order to interact with a CLI.

This diff also removes unused systemd files. These are no longer needed since
we've moved to a single service per container model.

Reviewed By: emakeev

Differential Revision: D15465016

fbshipit-source-id: cf9e4c7f1e6934849c6172a52440640dbd0aa6b4