Multiplex magmad network writes into configurator
Summary:
Add logic for multiplexing network related magmad writes into configurator. The multiplex happens on the obsidian handlers level so that it is easier to turn off the magmad writes later on.
For network creation, the http post operation will fail if the multiplex fails.
For network update, if the network already does not exist, it will just create a new network using the information passed into the handler.

Reviewed By: xjtian

Differential Revision: D15520750

fbshipit-source-id: 4cb53c443d68c8ce24316b66a696261275375630