Remove oai_sgw build target and udp/s11 tasks
Summary:
# UDP
The UDP task in OAI is used to communicate with the GTPV-U1 task. However, since moving to openflow for dataplane management, we no longer use this functionality. Let's remove it.

# S11
The S11 task in OAI is used to communicate with the SGW. However, since moving to embedded SGW, we no longer use this functionality. Let's remove it.

# OAI_SGW
We will no longer be supporting separate SGW (hasn't been tested for a long time -- probably doesn't work in its current state). Let's remove the build target from the codebase.

Reviewed By: ssanadhya

Differential Revision: D15558101

fbshipit-source-id: 74876bbf0c547cd498b1bc6ee64540a50181189b