Multiplex magmad gateway write into device and configurator (#306)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/306

Add logic for multiplexing gateway related magmad writes into configurator.
The multiplex happens on the obsidian handlers level so that it is easier to turn off the magmad writes later on.
Since configurator/device DBs and magmad DBs are not yet synched up, if a gateway is not already registered in configurator/device when update is called, it will fetch the necessary information (physicalID) from magmad and create a gateway with the update patched on.

Reviewed By: xjtian

Differential Revision: D15520701

fbshipit-source-id: 9f443274e9ddbefc9e500dabad0b3f0d655e0a9d