Multiplex writes into config into configurator
Summary:
Multiplexing obsidian config writes into configurator.
The logic cases on the format of config type to see whether to treat it as a network config or  an entity config. (if the type is prefixed by "network_", it's treated as a network config. Otherwise, it is treated as an entity config)

Additionally, since the configurator and config DBs have not been synched up, when multiplexing a config write into a network/entity that does not yet exist in the configurator DB, it will create a placeholder network/entity that only has the necessary information.

Reviewed By: xjtian

Differential Revision: D15498788

fbshipit-source-id: 9e68cd66b1238f866ec87c495d1504d6ed3474bf