Add orc8r gateway services to the CWAG with docker-compose
Summary:
This diff serves to add Magma services to the CWAG.
For now, it uses existing dockerfiles within the feg module.
Eventually this dockerfile will be moved to magma/orc8r/gateway
but for now this allows the CWAG to begin functioning as a gateway.

This diff also adds the necessary Golang requirements and service configs.

A diff stack on top of this one will add pipelined and sessiond services
to the CWAG as well.

Reviewed By: emakeev

Differential Revision: D15600099

fbshipit-source-id: ec4255e6a44de47aceac120c33c8af401e95c39e