Fix load service mconfig key error (#318)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/318

When the mconfig is initially loaded, the configs are loaded using different keys depending if it loads from the default directory or the override directory, loading with "configs_by_key" and "configsByKey" respectively. The load state is unknown at start so both keys need to be loaded when loading the mconfig.

Differential Revision: D15698062

fbshipit-source-id: 1ec61a2a27cdc37b1fb27cc65495c7fdb34e519f