Fix load service mconfig key error (#961)
Summary:
Pull Request resolved: https://github.com/facebookexternal/fbc/pull/961

Pull Request resolved: https://github.com/facebookincubator/magma/pull/318

When the mconfig is initially loaded, the configs are loaded using different keys depending if it loads from the default directory or the override directory, loading with "configs_by_key" and "configsByKey" respectively. The load state is unknown at start so both keys need to be loaded when loading the mconfig.

Reviewed By: xjtian

Differential Revision: D15698062

fbshipit-source-id: d6a18e71910b6831265d077028658f79532b1f79